requirements.txt
python-decouple==3.3
requests==2.24.0
loguru==0.5.3